modify cleanup_system.sh

+++ cleanup_system.sh
@@ -1,23 +1,34 @@
 #!/bin/bash
-#zeropadding the unused space
+#cleanup_system.sh
+#cleanup script for the system
+
+echo "zeropad the unused space"
 sudo dd if=/dev/zero of=zero bs=4k; sudo \rm zero
 
-#Remove old kernels
+echo "remove old kernels"
 sudo apt-get -y autoremove --purge
 
-#Remove apt cache
+echo "remove apt cache"
 sudo apt-get -y clean
 
-#Remove unnecessary kernels
+echo "remove unnecessary kernels"
 sudo purge-old-kernels
 
-#Remove logs
-rm ~/git/lin4neuro-bionic/*.log
+echo "remove logs"
+cd ~/git/lin4neuro-bionic
+ls | grep .log > /dev/null
+if [ $? -eq 0 ]; then
+    rm ~/git/lin4neuro-bionic/*.log
+fi
 
-#Remove downloads
-rm ~/Downloads/*
+echo "remove downloaded files"
+cd ~/Downloads
+fcount=$(ls | wc -l)
+if [ $fcount != 0 ]; then
+    rm -rf *
+fi
 
-#Remove history
+echo "remove history"
 export HISTSIZE=0; cd $HOME; rm .bash_history
 
 exit
